+++ src/nlua_naev.c
@@ -631,6 +631,7 @@ static int naevL_conf( lua_State *L )
    PUSH_BOOL( L, "fpu_except", conf.fpu_except );
    PUSH_STRING( L, "dev_data_dir", conf.dev_data_dir );
    PUSH_BOOL( L, "puzzle_skip", conf.puzzle_skip );
+   PUSH_BOOL( L, "disable_screen_shake", conf.disable_screen_shake );
    return 1;
 }
 #undef PUSH_STRING