+++ src/spfx.c
@@ -985,7 +985,7 @@ void spfx_shake( double mod )
    spfx_hapticRumble( mod );
 
    /* Create the shake. */
-   if ( shake_shader_pp_id == 0 )
+   if ( !conf.disable_screen_shake && ( shake_shader_pp_id == 0 ) )
       shake_shader_pp_id =
          render_postprocessAdd( &shake_shader, PP_LAYER_GAME, 99, 0 );
 }
@@ -1003,7 +1003,7 @@ void spfx_damage( double mod )
       MIN( SPFX_DAMAGE_MAX, damage_strength + SPFX_DAMAGE_MOD * mod );
 
    /* Create the damage. */
-   if ( damage_shader_pp_id == 0 )
+   if ( !conf.disable_screen_shake && ( damage_shader_pp_id == 0 ) )
       damage_shader_pp_id =
          render_postprocessAdd( &damage_shader, PP_LAYER_GUI, 98, 0 );
 }